Ajoute les cotisations forfaitaire de début d'activité
