+++ mon-entreprise/source/components/ui/Layout.css
@@ -38,7 +38,7 @@ section {
 }
 
 .ui__.side-block {
-	margin-top: 1rem;
+	padding-top: 1rem;
 	display: flex;
 	flex-direction: column;
 }
@@ -50,11 +50,13 @@ section {
 
 @media (min-width: 1200px) {
 	.ui__.side-block {
+		margin-top: -1rem;
+
 		flex-grow: 0;
 		flex-shrink: 1;
 		flex: 1;
 		position: sticky;
-		top: 1rem;
+		top: 0rem;
 	}
 }
