+++ mon-entreprise/source/locales/ui-en.yaml
@@ -179,6 +179,7 @@ Un seul associé: Only one partner
 Une idée&nbsp;?<1></1>Contactez-nous&nbsp;!: Any ideas?<1></1>Contact us!
 "Versement : ": "Payment : "
 Vie privée: Privacy
+Voir la fiche Urssaf: See the Urssaf sheet
 Voir la fiche de paie: See the pay slip
 Voir la répartition des cotisations: View contribution breakdown
 Voir le code source: See the source code
